add(Repository): add a bookings repository
